requirements.txt
chatlas[google]>=0.8.1
shiny>=1.4.0
python-dotenv>=1.1.0